requirements.txt
aiodns==4.0.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cocos2d==0.6.10
colorama==0.4.6
cryptography==46.0.5
distlib==0.3.8
distro==1.9.0
dj-database-url==3.1.2
Django==5.0.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==12.0.1
filelock==3.15.4
frozenlist==1.8.0
fsspec==2026.2.0
ghp-import==2.1.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
haidra_core==0.0.5
hf-xet==1.3.2
hiredis==3.3.0
horde_model_reference==2.1.5
horde_sdk==0.18.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mergedeep==1.3.4
mixer==7.2.2
mkdocs==1.6.1
mkdocs-get-deps==0.2.2
multidict==6.7.1
numpy==2.1.3
opencv-python==4.10.0.84
packaging==26.0
Panda3D==1.10.15
pathspec==1.0.4
pillow==11.0.0
platformdirs==4.2.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==5.0.1
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pygame==2.6.1
pyglet==1.5.29
Pygments==2.19.2
PyJWT==2.11.0
pymunk==6.9.0
PyOpenGL==3.1.7
pyparsing==3.3.2
pyscroll==2.31
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
PyTMX==3.32
PyYAML==6.0.3
pyyaml_env_tag==1.1
redis==7.3.0
replicate==1.0.7
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.1
StrEnum==0.4.15
stuntcat==0.2.4
tenacity==9.1.4
thorpy==2.1.10
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.26.3
watchdog==6.0.0
websockets==16.0
whitenoise==6.12.0
win32_setctime==1.2.0
yarl==1.23.0
